fixing rename for Nav2 staging
