+++ nav2_mppi_controller/CMakeLists.txt
@@ -88,6 +88,6 @@ ament_export_libraries(${libraries})
 ament_export_dependencies(${dependencies_pkgs})
 ament_export_include_directories(include)
 pluginlib_export_plugin_description_file(nav2_core mppic.xml)
-pluginlib_export_plugin_description_file(mppic critics.xml)
+pluginlib_export_plugin_description_file(nav2_mppi_controller critics.xml)
 
 ament_package()
