+++ nav2_mppi_controller/package.xml
@@ -36,7 +36,7 @@
   <export>
     <build_type>ament_cmake</build_type>
     <nav2_core plugin="${prefix}/mppic.xml" />
-    <mppic plugin="${prefix}/critics.xml" />
+    <nav2_mppi_controller plugin="${prefix}/critics.xml" />
   </export>
 
 </package>
